Module Bigstring_unix

String type based on Bigarray, for use in I/O and C-bindings, extending Core.Bigstring.

module Unix := Core_unix

include module type of struct include Core.Bigstring end

Types and exceptions

type t = (Base.Char.t, Bigarray.int8_unsigned_elt, Bigarray.c_layout) Bigarray.Array1.t

Type of bigstrings

include Ppx_compare_lib.Comparable.S with type t := t

include Ppx_compare_lib.Comparable.S__local with type t := t

include Ppx_quickcheck_runtime.Quickcheckable.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

type t_frozen = t

Type of bigstrings which support hashing. Note that mutation invalidates previous hashes.

include module type of Base_bigstring with type t := t and type t_frozen := t_frozen

Types and exceptions

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

include Ppx_compare_lib.Comparable.S__local with type t := t

include Ppx_compare_lib.Equal.S with type t := t

val equal : t -> t -> bool

include Ppx_compare_lib.Equal.S__local with type t := t

include Sexplib0.Sexpable.S with type t := t

include Sexplib0.Sexpable.Of_sexp with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

include Sexplib0.Sexpable.Sexp_of with type t := t

val sexp_of_t : t -> Sexplib0.Sexp.t

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t @@ portable

val globalize : t @ local read -> t @@ portable

val compare_t_frozen : t_frozen -> t_frozen -> Base.int

val compare_t_frozen__local : t_frozen @ local -> (t_frozen @ local -> Base.int) @ local

val equal_t_frozen : t_frozen -> t_frozen -> Base.bool

val equal_t_frozen__local : t_frozen @ local -> (t_frozen @ local -> Base.bool) @ local

val globalize_t_frozen : t_frozen @ local -> t_frozen

val hash_fold_t_frozen : Ppx_hash_lib.Std.Hash.state -> t_frozen -> Ppx_hash_lib.Std.Hash.state

val hash_t_frozen : t_frozen -> Ppx_hash_lib.Std.Hash.hash_value

val sexp_of_t_frozen : t_frozen -> Sexplib0.Sexp.t

val t_frozen_of_sexp : Sexplib0.Sexp.t -> t_frozen

val t_frozen_sexp_grammar : t_frozen Sexplib0.Sexp_grammar.t @@ portable

Creation and string conversion

val empty : t @@ portable

empty is a bigstring of length 0

val init : Base.int -> f:(Base.int -> Base.char) -> t @@ portable

init n ~f creates a bigstring t of length n, with t.{i} = f i.

val of_string : ?pos:Base.int -> ?len:Base.int -> Base.string @ local -> t @@ portable

of_string ?pos ?len str

returns
a new bigstring that is equivalent to the substring of length len in str starting at position pos.

parameter pos
default = 0

parameter len
default = String.length str - pos

val of_bytes : ?pos:Base.int -> ?len:Base.int -> Base.bytes @ local read -> t @@ portable

of_bytes ?pos ?len str

returns
a new bigstring that is equivalent to the subbytes of length len in str starting at position pos.

parameter pos
default = 0

parameter len
default = Bytes.length str - pos

val to_string : ?pos:Base.int -> ?len:Base.int -> t @ local read -> Base.string @@ portable

to_string ?pos ?len bstr

returns
a new string that is equivalent to the substring of length len in bstr starting at position pos.

parameter pos
default = 0

parameter len
default = length bstr - pos

raises Invalid_argument
if the string would exceed runtime limits.

val to_bytes : ?pos:Base.int -> ?len:Base.int -> t @ local read -> Base.bytes @@ portable

to_bytes ?pos ?len bstr

returns
a new byte sequence that is equivalent to the substring of length len in bstr starting at position pos.

parameter pos
default = 0

parameter len
default = length bstr - pos

raises Invalid_argument
if the bytes would exceed runtime limits.

val concat : ?sep:t @ local read -> (t Base.list @ local read -> t) @ local @@ portable

concat ?sep list returns the concatenation of list with sep in between each.

val unsafe_globalize_shared : t @ local -> t @@ portable

Like globalize, but does not copy the bigstring. Returns the input unchanged, as all bigstrings are allocated on the heap as custom blocks.

May be unsafe to hold on to the bigstring and access its storage if the source of the local value overwrites or deletes its storage later.

Checking

val check_args : loc:Base.string -> pos:Base.int -> len:Base.int -> t @ local immutable -> Base.unit @@ portable

check_args ~loc ~pos ~len bstr checks the position and length arguments pos and len for bigstrings bstr.

raises
Invalid_argument if these arguments are illegal for the given bigstring using loc to indicate the calling context.

val get_opt_len : t @ local immutable -> (pos:Base.int -> (Base.int Base.option @ local -> Base.int) @ local) @ local @@ portable

get_opt_len bstr ~pos opt_len

returns
the length of a subbigstring in bstr starting at position pos and given optional length opt_len. This function does not check the validity of its arguments. Use check_args for that purpose.

Accessors

val length : t @ local immutable -> Base.int @@ portable

length bstr

returns
the length of bigstring bstr.

val get : t @ read -> Base.int -> Base.char @@ portable

get t pos returns the character at pos

val unsafe_get : t @ read -> Base.int -> Base.char @@ portable

unsafe_get t pos returns the character at pos, without bounds checks.

val set : t -> Base.int -> Base.char -> Base.unit @@ portable

set t pos sets the character at pos

val unsafe_set : t -> Base.int -> Base.char -> Base.unit @@ portable

unsafe_set t pos sets the character at pos, without bounds checks.

val is_mmapped : t @ immutable -> Base.bool @@ portable

is_mmapped bstr

returns
whether the bigstring bstr is memory-mapped.

Blitting

blit ~src ?src_pos ?src_len ~dst ?dst_pos () blits src_len characters from src starting at position src_pos to dst at position dst_pos.

raises Invalid_argument
if the designated ranges are out of bounds.

include Base.Blit.S__read with type t := t

val blit : src:t @ local read -> (src_pos:int -> (dst:t @ local -> (dst_pos:int -> (len:int -> unit) @ local) @ local) @ local) @ local

val blito : src:t @ local read -> (?src_pos:int -> (?src_len:int -> (dst:t @ local -> (?dst_pos:int -> (unit -> unit) @ local) @ local) @ local) @ local) @ local

val unsafe_blit : src:t @ local read -> (src_pos:int -> (dst:t @ local -> (dst_pos:int -> (len:int -> unit) @ local) @ local) @ local) @ local

val sub : t @ local read -> (pos:int -> (len:int -> t) @ local) @ local

val subo : ?pos:int -> ?len:int -> t @ local read -> t

val copy : t @ local read -> t @@ portable

module To_string = Core.Bigstring.To_string

module From_string = Core.Bigstring.From_string

module To_bytes = Core.Bigstring.To_bytes

module From_bytes = Core.Bigstring.From_bytes

val memset : t @ local -> (pos:Base.int -> (len:Base.int -> (Base.char -> Base.unit) @ local) @ local) @ local @@ portable

memset t ~pos ~len c fills t with c within the range [pos, pos + len).

val unsafe_memset : t @ local -> (pos:Base.int -> (len:Base.int -> (Base.char -> Base.unit) @ local) @ local) @ local @@ portable

unsafe_memset t ~pos ~len c fills t with c within the range [pos, pos + len), without bounds checks.

Memcmp

val memcmp : t @ local read -> (pos1:Base.int -> (t @ local read -> (pos2:Base.int -> (len:Base.int -> Base.int) @ local) @ local) @ local) @ local @@ portable

memcmp t1 ~pos1 t2 ~pos2 ~len is like compare t1 t2 except performs the comparison on the subregions of t1 and t2 defined by pos1, pos2, and len.

val memcmp_bytes : t @ local read -> (pos1:Base.int -> (Base.Bytes.t @ local read -> (pos2:Base.int -> (len:Base.int -> Base.int) @ local) @ local) @ local) @ local @@ portable

memcmp_bytes, for efficient memcmp between Bigstring and Bytes data.

val memcmp_string : t @ local read -> (pos1:Base.int -> (Base.string @ local read -> (pos2:Base.int -> (len:Base.int -> Base.int) @ local) @ local) @ local) @ local @@ portable

memcmp_string, for efficient memcmp between Bigstring and string data.

val unsafe_strncmp : t @ read -> pos1:Base.int -> t @ read -> pos2:Base.int -> len:Base.int -> Base.int @@ portable

Compares up to len characters of two (potentially null-terminated) strings beginning at pos1 and pos2 of their respective Bigstrings. This function starts comparing the first character of each string. If they are equal to each other, it continues with the following pairs until the characters differ, until a terminating null-character is reached, or until len characters match in both strings, whichever happens first.

unsafe_strncmp does no bounds checking.

val find : ?pos:Base.int -> ?len:Base.int -> Base.char -> t @ local read -> Base.int Base.option @@ portable

find ?pos ?len char t returns Some i for the smallest i >= pos such that t.{i} = char, or None if there is no such i.

parameter pos
default = 0

parameter len
default = length bstr - pos

val rfind : ?pos:Base.int -> ?len:Base.int -> Base.char -> t @ local read -> Base.int Base.option @@ portable

rfind ?pos ?len char t returns Some i for the largest i >= pos such that t.{i} = char, or None if there is no such i.

parameter pos
default = 0

parameter len
default = length bstr - pos

val unsafe_find : t @ read -> Base.char -> pos:Base.int -> len:Base.int -> Base.int @@ portable

Same as find, but does no bounds checking, and returns a negative value instead of None if char is not found.

val unsafe_rfind : t @ read -> Base.char -> pos:Base.int -> len:Base.int -> Base.int @@ portable

Same as rfind, but does no bounds checking, and returns a negative value instead of None if char is not found.

val memmem : haystack:t @ local read -> (needle:t @ local read -> (?haystack_pos:Base.int -> (?haystack_len:Base.int -> (?needle_pos:Base.int -> (?needle_len:Base.int -> (Base.unit -> Base.int Base.option) @ local) @ local) @ local) @ local) @ local) @ local @@ portable

Search for the position of (a substring of) needle in (a substring of) haystack.

val unsafe_memmem : haystack:t @ read -> needle:t @ read -> haystack_pos:Base.int -> haystack_len:Base.int -> needle_pos:Base.int -> needle_len:Base.int -> Base.int @@ portable

As unsafe_find for memmem.

{2 Accessors for parsing binary values, analogous to [Binary_packing]}

   These are in [Bigstring] rather than a separate module because:

   1. Existing [Binary_packing] requires copies and does not work with [bigstring]s.
   2. The accessors rely on the implementation of [bigstring], and hence should change
   should the implementation of [bigstring] move away from [Bigarray].
   3. [Bigstring] already has some external C functions, so it didn't require many
   changes to the [jbuild] ^_^.

   In a departure from [Binary_packing], the naming conventions are chosen to be close to
   C99 stdint types, as it's a more standard description and it is somewhat useful in
   making compact macros for the implementations.  The accessor names contain endian-ness
   to allow for branch-free implementations

   <accessor>  ::= <unsafe><operation><type><endian>
   <unsafe>    ::= unsafe_ | ''
   <operation> ::= get_ | set_
   <type>      ::= int8 | uint8 | int16 | uint16 | int32 | uint32 | int64 | uint64
   <endian>    ::= _le | _be | ''

   The [unsafe_] prefix indicates that these functions do no bounds checking and silently
   truncate out-of-range numeric arguments.

val get_int8 : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val set_int8_exn : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val get_uint8 : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val set_uint8_exn : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_get_int8 : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_set_int8 : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_get_uint8 : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_set_uint8 : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

16-bit methods

val get_int16_le : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val get_int16_be : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val set_int16_le_exn : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val set_int16_be_exn : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_get_int16_le : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_get_int16_be : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_set_int16_le : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_set_int16_be : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val get_uint16_le : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val get_uint16_be : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val set_uint16_le_exn : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val set_uint16_be_exn : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_get_uint16_le : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_get_uint16_be : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_set_uint16_le : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_set_uint16_be : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

32-bit methods

val get_int32_le : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val get_int32_be : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val set_int32_le_exn : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val set_int32_be_exn : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_get_int32_le : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_get_int32_be : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_set_int32_le : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_set_int32_be : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val get_uint32_le : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val get_uint32_be : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val set_uint32_le_exn : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val set_uint32_be_exn : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_get_uint32_le : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_get_uint32_be : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_set_uint32_le : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_set_uint32_be : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

Similar to the usage in binary_packing, the below methods are treating the value being read (or written), as an ocaml immediate integer, as such it is actually 63 bits. If the user is confident that the range of values used in practice will not require 64-bit precision (i.e. Less than Max_Long), then we can avoid allocation and use an immediate. If the user is wrong, an exception will be thrown (for get).

64-bit signed values

val get_int64_le_exn : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val get_int64_be_exn : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val get_int64_le_trunc : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val get_int64_be_trunc : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val set_int64_le : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val set_int64_be : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_get_int64_le_exn : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_get_int64_be_exn : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_get_int64_le_trunc : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_get_int64_be_trunc : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_set_int64_le : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_set_int64_be : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

64-bit unsigned values

val get_uint64_be_exn : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val get_uint64_le_exn : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val set_uint64_le_exn : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val set_uint64_be_exn : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_get_uint64_be_exn : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_get_uint64_le_exn : t @ local read -> (pos:Base.int -> Base.int) @ local @@ portable

val unsafe_set_uint64_le : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

val unsafe_set_uint64_be : t @ local -> (pos:Base.int -> (Base.int -> Base.unit) @ local) @ local @@ portable

32-bit methods with full precision

val get_int32_t_le : t @ local read -> (pos:Base.int -> Base.Int32.t) @ local @@ portable

val get_int32_t_be : t @ local read -> (pos:Base.int -> Base.Int32.t) @ local @@ portable

val set_int32_t_le : t @ local -> (pos:Base.int -> (Base.Int32.t @ local -> Base.unit) @ local) @ local @@ portable

val set_int32_t_be : t @ local -> (pos:Base.int -> (Base.Int32.t @ local -> Base.unit) @ local) @ local @@ portable

val unsafe_get_int32_t_le : t @ local read -> (pos:Base.int -> Base.Int32.t) @ local @@ portable

val unsafe_get_int32_t_be : t @ local read -> (pos:Base.int -> Base.Int32.t) @ local @@ portable

val unsafe_set_int32_t_le : t @ local -> (pos:Base.int -> (Base.Int32.t @ local -> Base.unit) @ local) @ local @@ portable

val unsafe_set_int32_t_be : t @ local -> (pos:Base.int -> (Base.Int32.t @ local -> Base.unit) @ local) @ local @@ portable

64-bit methods with full precision

val get_int64_t_le : t @ local read -> (pos:Base.int -> Base.Int64.t) @ local @@ portable

val get_int64_t_be : t @ local read -> (pos:Base.int -> Base.Int64.t) @ local @@ portable

val set_int64_t_le : t @ local -> (pos:Base.int -> (Base.Int64.t @ local -> Base.unit) @ local) @ local @@ portable

val set_int64_t_be : t @ local -> (pos:Base.int -> (Base.Int64.t @ local -> Base.unit) @ local) @ local @@ portable

val unsafe_get_int64_t_le : t @ local read -> (pos:Base.int -> Base.Int64.t) @ local @@ portable

val unsafe_get_int64_t_be : t @ local read -> (pos:Base.int -> Base.Int64.t) @ local @@ portable

val unsafe_set_int64_t_le : t @ local -> (pos:Base.int -> (Base.Int64.t @ local -> Base.unit) @ local) @ local @@ portable

val unsafe_set_int64_t_be : t @ local -> (pos:Base.int -> (Base.Int64.t @ local -> Base.unit) @ local) @ local @@ portable

String methods

These are alternatives to to_string that follow the conventions of the int accessors, and in particular avoid optional arguments.

val get_string : t @ local read -> (pos:Base.int -> (len:Base.int -> Base.string) @ local) @ local @@ portable

val unsafe_get_string : t @ local read -> (pos:Base.int -> (len:Base.int -> Base.string) @ local) @ local @@ portable

module Local = Core.Bigstring.Local

module Int_repr = Core.Bigstring.Int_repr

module Private = Core.Bigstring.Private

include Core.Hexdump.S with type t := t

module Hexdump = Core.Bigstring.Hexdump

Creation and string conversion

val create : Base.Int.t -> t @@ portable

create length

returns
a new bigstring having length. Content is undefined.

val sub_shared : ?pos:Base.Int.t -> ?len:Base.Int.t -> t -> t @@ portable

sub_shared ?pos ?len bstr

returns
the sub-bigstring in bstr that starts at position pos and has length len. The sub-bigstring shares the same memory region, i.e. modifying it will modify the original bigstring. Holding on to the sub-bigstring will also keep the (usually bigger) original one around.

parameter pos
default = 0

parameter len
default = Bigstring.length bstr - pos

val sub_shared_local : ?pos:Base.Int.t -> ?len:Base.Int.t -> t @ local -> t @ local @@ portable

Like sub_shared, for local input and output.

The result is allocated on the global heap, even if built with a compiler supporting stack allocation. At least as of 2024-08, custom blocks with finalizers cannot be allocated on the local heap.

val unsafe_sub_shared_of_local : ?pos:Base.Int.t -> ?len:Base.Int.t -> t @ local -> t @@ portable

Like sub_shared, for local input and global output.

Creates a global bigstring sharing the same storage as the local input. This is unsafe if the input's storage will be destroyed or overwritten after t's local scope ends. Only use this function if you know the input's underlying storage is safe to use beyond the lifetime of the input t.

Reading/writing bin-prot

These functions write the "size-prefixed" bin-prot format that is used by, e.g., async's Writer.write_bin_prot, Reader.read_bin_prot and Unpack_buffer.Unpack_one.create_bin_prot.

val write_bin_prot : t -> ?pos:Base.Int.t -> 'a Bin_prot.Type_class.writer -> 'a -> Base.Int.t @@ portable

write_bin_prot t writer a writes a to t starting at pos, and returns the index in t immediately after the last byte written. It raises if pos < 0 or if a doesn't fit in t.

val write_bin_prot_known_size : t -> ?pos:Base.Int.t -> 'a Bin_prot.Write.writer -> size:Base.Int.t -> 'a -> Base.Int.t @@ portable

Same as write_bin_prot, with the difference that size is pre-computed by the caller. size is assumed to be the result of calling the bin prot sizer on the value being written.

val read_bin_prot : t -> ?pos:Base.Int.t -> ?len:Base.Int.t -> 'a Bin_prot.Type_class.reader -> ('a * Base.Int.t) Core.Or_error.t @@ portable

The read_bin_prot* functions read from the region of t starting at pos of length len. They return the index in t immediately after the last byte read. They raise if pos and len don't describe a region of t.

val read_bin_prot_verbose_errors : t -> ?pos:Base.Int.t -> ?len:Base.Int.t -> 'a Bin_prot.Type_class.reader -> [ `Invalid_data of Core.Error.t | `Not_enough_data | `Ok of 'a * Base.Int.t ] @@ portable

Destruction

val unsafe_destroy : t -> Base.Unit.t @@ portable

unsafe_destroy bstr destroys the bigstring by deallocating its associated data or, if memory-mapped, unmapping the corresponding file, and setting all dimensions to zero. This effectively frees the associated memory or address-space resources instantaneously. This feature helps reclaim the resources sooner than they are automatically reclaimed by the GC.

This operation is safe unless you have passed the bigstring to another thread that is performing operations on it at the same time. Access to the bigstring after this operation will yield array bounds exceptions.

raises Failure
if the bigstring has already been deallocated (or deemed "external", which is treated equivalently), or if it has proxies, i.e. other bigstrings referring to the same data.

val unsafe_destroy_and_resize : t -> len:Base.Int.t -> t @@ portable

unsafe_destroy_and_resize bstr ~len reallocates the memory backing bstr and returns a new bigstring that starts at position 0 and has length len. If len is greater than length bstr then the newly allocated memory will not be initialized.

Similar to unsafe_destroy, this operation is safe unless you have passed the bigstring to another thread that is performing operations on it at the same time. Access to bstr after this operation will yield array bounds exceptions.

raises Failure
if the bigstring has already been deallocated (or deemed "external", which is treated equivalently), if it is backed by a memory map, or if it has proxies, i.e. other bigstrings referring to the same data.

val get_tail_padded_fixed_string : padding:Base.Char.t -> t @ local -> (pos:Base.Int.t -> (len:Base.Int.t -> (Base.Unit.t -> Base.String.t) @ local) @ local) @ local @@ portable

Similar to Binary_packing.unpack_tail_padded_fixed_string and .pack_tail_padded_fixed_string.

val get_tail_padded_fixed_string_local : padding:Base.Char.t -> t @ local -> (pos:Base.Int.t -> (len:Base.Int.t -> (Base.Unit.t -> Base.String.t @ local) @ local) @ local) @ local @@ portable

val set_tail_padded_fixed_string : padding:Base.Char.t -> t @ local -> (pos:Base.Int.t -> (len:Base.Int.t -> (Base.String.t @ local -> Base.Unit.t) @ local) @ local) @ local @@ portable

val get_head_padded_fixed_string : padding:Base.Char.t -> t @ local -> (pos:Base.Int.t -> (len:Base.Int.t -> (Base.Unit.t -> Base.String.t) @ local) @ local) @ local @@ portable

val get_head_padded_fixed_string_local : padding:Base.Char.t -> t @ local -> (pos:Base.Int.t -> (len:Base.Int.t -> (Base.Unit.t -> Base.String.t @ local) @ local) @ local) @ local @@ portable

val set_head_padded_fixed_string : padding:Base.Char.t -> t @ local -> (pos:Base.Int.t -> (len:Base.Int.t -> (Base.String.t @ local -> Base.Unit.t) @ local) @ local) @ local @@ portable

module Unstable = Core.Bigstring.Unstable

module Stable = Core.Bigstring.Stable

exception IOError of int * exn

Type of I/O errors.

In IOError (n, exn), n is the number of bytes successfully read/written before the error and exn is the exception that occurred (e.g., Unix_error, End_of_file)

Input functions

val read : ?min_len:int -> Unix.File_descr.t -> ?pos:int -> ?len:int -> t @ local -> int @@ portable

read ?min_len fd ?pos ?len bstr reads at least min_len (must be >= 0) and at most len (must be >= min_len) bytes from file descriptor fd, and writes them to bigstring bstr starting at position pos. Returns the number of bytes actually read.

read returns zero only if len = 0. If len > 0 and there's nothing left to read, read raises to indicate EOF even if min_len = 0.

NOTE: Even if len is zero, there may still be errors when reading from the descriptor!

Raises Invalid_argument if the designated ranges are out of bounds. Raises IOError in the case of input errors, or on EOF if the minimum length could not be read.

val really_read : Unix.File_descr.t -> ?pos:int -> ?len:int -> t @ local -> unit @@ portable

really_read fd ?pos ?len bstr reads len bytes from file descriptor fd, and writes them to bigstring bstr starting at position pos.

Raises Invalid_argument if the designated range is out of bounds. Raises IOError in the case of input errors, or on EOF.

val really_recv : Unix.File_descr.t -> ?pos:int -> ?len:int -> t @ local -> unit @@ portable

really_recv sock ?pos ?len bstr receives len bytes from socket sock, and writes them to bigstring bstr starting at position pos. If len is zero, the function returns immediately without performing the underlying system call.

Raises Invalid_argument if the designated range is out of bounds. Raises IOError in the case of input errors, or on EOF.

val recv_peek_assume_fd_is_nonblocking : Unix.File_descr.t -> ?pos:int -> len:int -> t @ local -> int @@ portable

recv_peek_assume_fd_is_nonblocking sock ?pos ~len bstr peeks len bytes from socket sock, and writes them to bigstring bstr starting at position pos. If len is zero, the function returns immediately without performing the underlying system call.

Raises Invalid_argument if the designated range is out of bounds. Raises Unix_error in the case of input errors

val recvfrom_assume_fd_is_nonblocking : Unix.File_descr.t -> ?pos:int -> ?len:int -> t @ local -> int * Unix.sockaddr @@ portable

recvfrom_assume_fd_is_nonblocking sock ?pos ?len bstr reads up to len bytes into bigstring bstr starting at position pos from socket sock without yielding to other OCaml-threads.

Returns the number of bytes actually read and the socket address of the client.

Raises Unix_error in the case of input errors. Raises Invalid_argument if the designated range is out of bounds.

val read_assume_fd_is_nonblocking : Unix.File_descr.t -> ?pos:int -> ?len:int -> t @ local -> Unix.Syscall_result.Int.t @@ portable

read_assume_fd_is_nonblocking fd ?pos ?len bstr reads up to len bytes into bigstring bstr starting at position pos from file descriptor fd without yielding to other OCaml-threads. Returns the number of bytes actually read.

Raises Invalid_argument if the designated range is out of bounds.

val pread : ?min_len:int -> Unix.File_descr.t -> offset:int -> ?pos:int -> ?len:int -> t @ local -> int @@ portable

Like read but uses pread to read from the given offset in the file.

val really_pread : Unix.File_descr.t -> offset:int -> ?pos:int -> ?len:int -> t @ local -> unit @@ portable

Like really_read but uses pread to read from the given offset in the file.

val pread_assume_fd_is_nonblocking : Unix.File_descr.t -> offset:int -> ?pos:int -> ?len:int -> t @ local -> int @@ portable

pread_assume_fd_is_nonblocking fd ~offset ?pos ?len bstr reads up to len bytes from file descriptor fd at offset offset, and writes them to bigstring bstr starting at position pos. The fd must be capable of seeking, and the current file offset used for a regular read() is unchanged. Please see man pread for more information. Returns the number of bytes actually read.

Raises Invalid_argument if the designated range is out of bounds. Raises Unix_error in the case of input errors.

val input : ?min_len:int -> Core.In_channel.t -> ?pos:int -> ?len:int -> t @ local -> int @@ portable

input ?min_len ic ?pos ?len bstr tries to read len bytes (guarantees to read at least min_len bytes, which must be >= 0 and <= len), if possible, before returning, from input channel ic, and writes them to bigstring bstr starting at position pos. Returns the number of bytes actually read.

NOTE: Even if len is zero, there may still be errors when reading from the descriptor, which will be done if the internal buffer is empty!

NOTE: If at least len characters are available in the input channel buffer and if len is not zero, data will only be fetched from the channel buffer. Otherwise data will be read until at least min_len characters are available.

Raises Invalid_argument if the designated range is out of bounds. Raises IOError in the case of input errors, or on premature EOF.

val really_input : Core.In_channel.t -> ?pos:int -> ?len:int -> t @ local -> unit @@ portable

really_input ic ?pos ?len bstr reads exactly len bytes from input channel ic, and writes them to bigstring bstr starting at position pos.

Raises Invalid_argument if the designated range is out of bounds. Raises IOError in the case of input errors, or on premature EOF.

Output functions

val really_write : Unix.File_descr.t -> ?pos:int -> ?len:int -> t @ local shared -> unit @@ portable

really_write fd ?pos ?len bstr writes len bytes in bigstring bstr starting at position pos to file descriptor fd.

Raises Invalid_argument if the designated range is out of bounds. Raises IOError in the case of output errors.

val really_send_no_sigpipe : (Unix.File_descr.t -> ?pos:int -> ?len:int -> t @ local shared -> unit) Core.Or_error.t @@ portable

really_send_no_sigpipe sock ?pos ?len bstr sends len bytes in bigstring bstr starting at position pos to socket sock without blocking and ignoring SIGPIPE.

Raises Invalid_argument if the designated range is out of bounds. Raises IOError in the case of output errors.

really_send_no_sigpipe is not implemented on some platforms, in which case it returns an Error value indicating that it is unimplemented.

val send_nonblocking_no_sigpipe : (Unix.File_descr.t -> ?pos:int -> ?len:int -> t @ local shared -> Unix.Syscall_result.Int.t) Core.Or_error.t @@ portable

send_nonblocking_no_sigpipe sock ?pos ?len bstr tries to send len bytes in bigstring bstr starting at position pos to socket sock. Returns bytes_written.

Raises Invalid_argument if the designated range is out of bounds.

val sendto_nonblocking_no_sigpipe : (Unix.File_descr.t -> ?pos:int -> ?len:int -> t @ local shared -> (Unix.sockaddr -> Unix.Syscall_result.Int.t) @ local) Core.Or_error.t @@ portable

sendto_nonblocking_no_sigpipe sock ?pos ?len bstr sockaddr tries to send len bytes in bigstring bstr starting at position pos to socket sock using address addr. Returns bytes_written.

Raises Invalid_argument if the designated range is out of bounds.

val write : Unix.File_descr.t -> ?pos:int -> ?len:int -> t @ local shared -> int @@ portable

write fd ?pos ?len bstr writes len bytes in bigstring bstr starting at position pos to file descriptor fd. Returns the number of bytes actually written.

Raises Invalid_argument if the designated range is out of bounds. Raises Unix_error in the case of output errors.

val pwrite_assume_fd_is_nonblocking : Unix.File_descr.t -> offset:int -> ?pos:int -> ?len:int -> t @ local shared -> int @@ portable

pwrite_assume_fd_is_nonblocking fd ~offset ?pos ?len bstr writes up to len bytes of bigstring bstr starting at position pos to file descriptor fd at position offset. The fd must be capable of seeking, and the current file offset used for non-positional read()/write() calls is unchanged. Returns the number of bytes written.

Raises Invalid_argument if the designated range is out of bounds. Raises Unix_error in the case of output errors.

val write_assume_fd_is_nonblocking : Unix.File_descr.t -> ?pos:int -> ?len:int -> t @ local shared -> int @@ portable

write_assume_fd_is_nonblocking fd ?pos ?len bstr writes len bytes in bigstring bstr starting at position pos to file descriptor fd without yielding to other OCaml-threads. Returns the number of bytes actually written.

Raises Invalid_argument if the designated range is out of bounds. Raises Unix_error in the case of output errors.

val writev : Unix.File_descr.t -> ?count:int -> t Unix.IOVec.t array @ shared -> int @@ portable

writev fd ?count iovecs writes count iovecs of bigstrings to file descriptor fd. Returns the number of bytes written.

Raises Invalid_argument if count is out of range. Raises Unix_error in the case of output errors.

val writev_assume_fd_is_nonblocking : Unix.File_descr.t -> ?count:int -> t Unix.IOVec.t array @ shared -> int @@ portable

writev_assume_fd_is_nonblocking fd ?count iovecs writes count iovecs of bigstrings to file descriptor fd without yielding to other OCaml-threads. Returns the number of bytes actually written.

Raises Invalid_argument if the designated range is out of bounds. Raises Unix_error in the case of output errors.

val recvmmsg_assume_fd_is_nonblocking : (Unix.File_descr.t -> ?count:int -> ?srcs:Unix.sockaddr array -> t Unix.IOVec.t array -> lens:int array -> int) Core.Or_error.t @@ portable

recvmmsg_assume_fd_is_nonblocking fd iovecs ~count ~lens receives up to count messages into iovecs from file descriptor fd without yielding to other OCaml threads. If ~count is supplied, it must be that 0 <= count <= Array.length iovecs. If ~srcs is supplied, saves the source addresses for corresponding received messages there. If supplied, Array.length srcs must be >= count. Saves the lengths of the received messages in lens. It is required that Array.length lens >= count.

If an IOVec isn't long enough for its corresponding message, excess bytes may be discarded, depending on the type of socket the message is received from. While the recvmmsg system call itself does return details of such truncation, etc., those details are not (yet) passed through this interface.

See "recvmmsg(2)" re. the underlying system call.

Returns the number of messages actually read, or a negative number to indicate EWOULDBLOCK or EAGAIN. This is a compromise to mitigate the exception overhead for what ends up being a very common result with our use of recvmmsg.

Raises Invalid_argument if the designated range is out of bounds. Raises Unix_error in the case of output errors.

val unsafe_recvmmsg_assume_fd_is_nonblocking : (Unix.File_descr.t -> t Unix.IOVec.t array -> int -> Unix.sockaddr array option -> int array -> int) Core.Or_error.t @@ portable

val sendmsg_nonblocking_no_sigpipe : (Unix.File_descr.t -> ?count:int -> t Unix.IOVec.t array @ shared -> int option) Core.Or_error.t @@ portable

sendmsg_nonblocking_no_sigpipe sock ?count iovecs sends count iovecs of bigstrings to socket sock. Returns Some bytes_written, or None if the operation would have blocked. This system call will not cause signal SIGPIPE if an attempt is made to write to a socket that was closed by the other side.

Raises Invalid_argument if count is out of range. Raises Unix_error in the case of output errors.

val output : ?min_len:int -> Core.Out_channel.t -> ?pos:int -> ?len:int -> t @ local shared -> int @@ portable

output ?min_len oc ?pos ?len bstr tries to output len bytes (guarantees to write at least min_len bytes, which must be >= 0), if possible, before returning, from bigstring bstr starting at position pos to output channel oc. Returns the number of bytes actually written.

NOTE: You may need to flush oc to make sure that the data is actually sent.

NOTE: If len characters fit into the channel buffer completely, they will be buffered. Otherwise writes will be attempted until at least min_len characters have been sent.

Raises Invalid_argument if the designated range is out of bounds.

Raises IOError in the case of output errors. The IOError argument counting the number of successful bytes includes those that have been transferred to the channel buffer before the error.

val really_output : Core.Out_channel.t -> ?pos:int -> ?len:int -> t @ local shared -> unit @@ portable

really_output oc ?pos ?len bstr outputs exactly len bytes from bigstring bstr starting at position pos to output channel oc.

Raises Invalid_argument if the designated range is out of bounds.

Raises IOError in the case of output errors. The IOError argument counting the number of successful bytes includes those that have been transferred to the channel buffer before the error.

Unsafe functions

val unsafe_read_assume_fd_is_nonblocking : Unix.File_descr.t -> pos:int -> len:int -> t @ local -> Unix.Syscall_result.Int.t @@ portable

unsafe_read_assume_fd_is_nonblocking fd ~pos ~len bstr is similar to Bigstring.read_assume_fd_is_nonblocking, but does not perform any bounds checks. Will crash on bounds errors!

val unsafe_write : Unix.File_descr.t -> pos:int -> len:int -> t @ local shared -> int @@ portable

unsafe_write fd ~pos ~len bstr is similar to Bigstring.write, but does not perform any bounds checks. Will crash on bounds errors!

val unsafe_write_assume_fd_is_nonblocking : Unix.File_descr.t -> pos:int -> len:int -> t @ local shared -> int @@ portable

unsafe_write_assume_fd_is_nonblocking fd ~pos ~len bstr is similar to Bigstring.write_assume_fd_is_nonblocking, but does not perform any bounds checks. Will crash on bounds errors!

val unsafe_read : min_len:int -> Unix.File_descr.t -> pos:int -> len:int -> t @ local -> int @@ portable

unsafe_read ~min_len fd ~pos ~len bstr is similar to Bigstring.read, but does not perform any bounds checks. Will crash on bounds errors!

val unsafe_really_recv : Unix.File_descr.t -> pos:int -> len:int -> t @ local -> unit @@ portable

unsafe_really_recv sock ~pos ~len bstr is similar to Bigstring.really_recv, but does not perform any bounds checks. Will crash on bounds errors!

val unsafe_really_write : Unix.File_descr.t -> pos:int -> len:int -> t @ local shared -> unit @@ portable

unsafe_really_write fd ~pos ~len bstr is similar to Bigstring.write, but does not perform any bounds checks. Will crash on bounds errors!

val unsafe_really_send_no_sigpipe : (Unix.File_descr.t -> pos:int -> len:int -> t @ local shared -> unit) Core.Or_error.t @@ portable

unsafe_really_send_no_sigpipe sock ~pos ~len bstr is similar to Bigstring.send, but does not perform any bounds checks. Will crash on bounds errors!

val unsafe_send_nonblocking_no_sigpipe : (Unix.File_descr.t -> pos:int -> len:int -> t @ local shared -> Unix.Syscall_result.Int.t) Core.Or_error.t @@ portable

unsafe_send_nonblocking_no_sigpipe sock ~pos ~len bstr is similar to Bigstring.send_nonblocking_no_sigpipe, but does not perform any bounds checks. Will crash on bounds errors!

val unsafe_writev : Unix.File_descr.t -> t Unix.IOVec.t array @ shared -> int -> int @@ portable

unsafe_writev fd iovecs count is similar to Bigstring.writev, but does not perform any bounds checks. Will crash on bounds errors!

val unsafe_sendmsg_nonblocking_no_sigpipe : (Unix.File_descr.t -> t Unix.IOVec.t array @ shared -> int -> int option) Core.Or_error.t @@ portable

unsafe_sendmsg_nonblocking_no_sigpipe fd iovecs count is similar to Bigstring.sendmsg_nonblocking_no_sigpipe, but does not perform any bounds checks. Will crash on bounds errors!

val unsafe_input : min_len:int -> Core.In_channel.t -> pos:int -> len:int -> t @ local -> int @@ portable

unsafe_input ~min_len ic ~pos ~len bstr is similar to Bigstring.input, but does not perform any bounds checks. Will crash on bounds errors!

val unsafe_output : min_len:int -> Core.Out_channel.t -> pos:int -> len:int -> t @ local shared -> int @@ portable

unsafe_output ~min_len oc ~pos ~len bstr is similar to Bigstring.output, but does not perform any bounds checks. Will crash on bounds errors!

Memory mapping

val map_file : shared:bool -> Unix.File_descr.t -> int -> t @@ portable

map_file shared fd n memory-maps n characters of the data associated with descriptor fd to a bigstring. Iff shared is true, all changes to the bigstring will be reflected in the file.

Users must keep in mind that operations on the resulting bigstring may result in disk operations which block the runtime. This is true for pure OCaml operations (such as t.{1} <- 1), and for calls to blit. While some I/O operations may release the OCaml lock, users should not expect this to be done for all operations on a bigstring returned from map_file.